# Optimize memory-allocations a bit here and there