--- a/griblib/gribtest/Makefile
+++ b/griblib/gribtest/Makefile
@@ -1,3 +1,5 @@
+.PHONY: default benchmark
+
 default:
 	go test -v