--- a/griblib/gribtest/performance_test.go
+++ b/griblib/gribtest/performance_test.go
@@ -10,12 +10,11 @@ import (
 func BenchmarkReadMessages(b *testing.B) {
 	// run the Fib function b.N times
 
-	f, err := os.Open("../integrationtestdata/template5_3.grib2")
-	if err != nil {
-		b.Fatalf("Could not open test-file %v", err)
-	}
-
 	for n := 0; n < b.N; n++ {
+		f, err := os.Open("../integrationtestdata/template5_3.grib2")
+		if err != nil {
+			b.Fatalf("Could not open test-file %v", err)
+		}
 		griblib.ReadMessages(f)
 	}
 }