@nilsmagnus 2019-02-07
What do you think @landru29, are these fixes correct?  It seems that some more optimizations can be done, but I need to figure out a clever way to do that.